requirements.txt
APScheduler==3.7.0
certifi==2020.12.5
click==7.1.2
Flask==1.1.2
flask-talisman==0.7.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
pytz==2021.1
selenium==3.141.0
sentry-sdk==0.20.1
six==1.15.0
tzlocal==2.1
urllib3==1.26.2
Werkzeug==1.0.1
